New version of the website using cirrus.css
This version simplifies the initial implementation
using bootstrap, and the landing page.

A similar style has been choosen to avoid having
a totally different experience.

The urls are maintained to avoid broken links.

A more smart mechanism for adding pages was implemented
and now we use meta information from the top of the
markdown pages.

New pages for the community initiatives were added.

The audio files from the podcast are not tracked in the repo